+++ apps/server/src/modules/files/job/delete-files.console.spec.ts
@@ -1,10 +1,10 @@
 import moment from 'moment';
 import { Test, TestingModule } from '@nestjs/testing';
 import { FileStorageAdapter } from '@shared/infra/filestorage';
-import { LoggerModule } from '../../../core/logger/logger.module';
+import { FilesRepo } from '@shared/repo';
+import { LoggerModule } from '@src/core/logger';
 import { DeleteFilesConsole } from './delete-files.console';
 import { DeleteFilesUc } from '../uc';
-import { FilesRepo } from '../repo';
 
 describe('DeleteFilesController', () => {
 	let console: DeleteFilesConsole;