+++ apps/server/src/shared/repo/index.ts
@@ -6,6 +6,7 @@
 
 export * from './course';
 export * from './coursegroup';
+export * from './dashboard';
 export * from './files';
 export * from './lesson';
 export * from './news';