+++ apps/server/src/modules/files/uc/delete-files.uc.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { Logger } from '@src/core/logger/logger.service';
-import { FilesRepo } from '../repo';
+import { FilesRepo } from '@shared/repo';
 
 @Injectable()
 export class DeleteFilesUc {