+++ apps/server/src/shared/domain/entity/submission.entity.spec.ts
@@ -2,40 +2,50 @@ import { Collection } from '@mikro-orm/core';
 import { userFactory, taskFactory, submissionFactory, fileFactory, setupEntities } from '@shared/testing';
 import { File } from './file.entity';
 
-const buildSubmission = () => {
-	const student = userFactory.build();
-	const task = taskFactory.build();
-	const submission = submissionFactory.build({ task, student });
-	return submission;
-};
-
 describe('Submission entity', () => {
 	beforeAll(async () => {
 		await setupEntities();
 	});
 
-	it('should be graded if grade percentage is set', () => {
-		const submission = buildSubmission();
-		submission.grade = 50;
-		expect(submission.isGraded()).toEqual(true);
+	describe('isGraded', () => {
+		it('should be graded if grade percentage is set', () => {
+			const student = userFactory.build();
+			const task = taskFactory.build();
+			const submission = submissionFactory.build({ task, student, grade: 50 });
+
+			expect(submission.isGraded()).toEqual(true);
+		});
+
+		it('should be graded if grade comment is set', () => {
+			const student = userFactory.build();
+			const task = taskFactory.build();
+			const submission = submissionFactory.build({ task, student, gradeComment: 'well done!' });
+
+			expect(submission.isGraded()).toEqual(true);
+		});
+
+		it('should be graded if grade grade files have been associated', () => {
+			const student = userFactory.build();
+			const task = taskFactory.build();
+			const teacher = userFactory.build();
+			const file = fileFactory.build({ creator: teacher });
+			const submission = submissionFactory.build({ task, student });
+			submission.gradeFiles = new Collection<File>(submission, [file]);
+
+			expect(submission.isGraded()).toEqual(true);
+		});
 	});
 
-	it('should be graded if grade comment is set', () => {
-		const submission = buildSubmission();
-		submission.gradeComment = 'well done!';
-		expect(submission.isGraded()).toEqual(true);
-	});
+	describe('getStudentId', () => {
+		it('should return the id of the student', () => {
+			const student = userFactory.build();
+			student.id = '0123456789ab';
+			const task = taskFactory.build();
+			const submission = submissionFactory.build({ task, student });
 
-	it('should be graded if grade grade files have been associated', () => {
-		const submission = buildSubmission();
-		const teacher = userFactory.build();
-		const file = fileFactory.build({ creator: teacher });
-		submission.gradeFiles = new Collection<File>(submission, [file]);
-		expect(submission.isGraded()).toEqual(true);
-	});
+			const result = submission.getStudentId();
 
-	it('should not be graded if none of the grade properties are set', () => {
-		const submission = buildSubmission();
-		expect(submission.isGraded()).toEqual(false);
+			expect(result).toEqual(student.id);
+		});
 	});
 });