+++ test/services/school/index.test.js
@@ -219,6 +219,38 @@ describe('school service', () => {
 
 			expect(usersSchoolUpdated.systems).to.not.be.empty;
 		});
+
+		it('should sync federalState in systems', async () => {
+			const BB = '0000b186816abba584714c53';
+			const BW = '0000b186816abba584714c50';
+			const school = await testObjects.createTestSchool({ federalState: BB });
+			const user = await testObjects.createTestUser({
+				schoolId: school._id,
+				roles: ['superhero', 'administrator'],
+			});
+			const params = await testObjects.generateRequestParamsFromUser(user);
+
+			const ldapSystemConfig = {
+				ldapConfig: {
+					provider: 'general',
+					federalState: BB,
+					active: true,
+				},
+				type: 'ldap',
+			};
+			let system = await testObjects.createTestSystem(ldapSystemConfig);
+
+			school.systems = [system._id];
+			await app.service('/schools').update(school._id, school, params);
+			system = await app.service('systems').get(system._id, params);
+			expect(system.ldapConfig.federalState.toString()).to.equal(BB);
+
+			school.federalState = BW;
+			await app.service('/schools').update(school._id, school, params);
+
+			system = await app.service('systems').get(system._id, params);
+			expect(system.ldapConfig.federalState.toString()).to.equal(BW);
+		});
 	});
 
 	describe('patch schools', () => {
@@ -507,6 +539,34 @@ describe('school service', () => {
 			const usersSchoolUpdated = await app.service('schools').get(school._id, params);
 
 			expect(usersSchoolUpdated.systems).to.not.be.empty;
+		});
+
+		it('should sync federalSatate in systems', async () => {
+			const BB = '0000b186816abba584714c53';
+			const BW = '0000b186816abba584714c50';
+			const school = await testObjects.createTestSchool({ federalState: BB });
+			const user = await testObjects.createTestUser({
+				schoolId: school._id,
+				roles: ['superhero', 'administrator'],
+			});
+			const params = await testObjects.generateRequestParamsFromUser(user);
+			const ldapSystemConfig = {
+				ldapConfig: {
+					provider: 'general',
+					federalState: BB,
+					active: true,
+				},
+				type: 'ldap',
+			};
+			let system = await testObjects.createTestSystem(ldapSystemConfig);
+			await app.service('/schools').patch(school._id, { systems: [system._id] }, params);
+			system = await app.service('systems').get(system._id, params);
+
+			expect(system.ldapConfig.federalState.toString()).to.equal(BB);
+
+			await app.service('/schools').patch(school._id, { federalState: BW }, params);
+			system = await app.service('systems').get(system._id, params);
+			expect(system.ldapConfig.federalState.toString()).to.equal(BW);
 		});
 	});
 });